🔥 refactor(`boot/platform`): Remove unused dependencies and reorganize package structure. 🚚 refactor(`boot/platform/src/main/java/com/plate/boot/inventory`): Move and rename packages to `inventory`. ➕ refactor(`boot/platform/src/main/resources/schema-postgres.sql`): Add new `event_publication` table and indexes. 📝 docs(`boot/platform/src/main/java/com/plate/boot/BootApplication.java`): Update documentation for the `BootApplication` class. 🗑️ chore(`build.gradle`): Remove `springModulith` related configurations.
