+++ boot/platform/src/main/java/com/plate/boot/BootApplication.java
@@ -4,7 +4,18 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
- * @author <a href="https://github.com/vnobo">Alex Bob</a>
+ * Entry point for the Spring Boot application.
+ * <p>
+ * This class serves as the primary class to launch the Spring Boot application.
+ * It is annotated with `@SpringBootApplication`, which is a convenience annotation
+ * that includes `@Configuration`, `@EnableAutoConfiguration`, and `@ComponentScan`.
+ * These annotations together configure the application context, enable auto-configuration
+ * of Spring features, and scan for Spring components in the package where this class is located
+ * and its sub-packages.
+ * <p>
+ * The `main` method initiates the application's run process by calling
+ * `SpringApplication.run(BootApplication.class, args)`, where `args` are command-line arguments
+ * passed to the application, if any.
  */
 @SpringBootApplication
 public class BootApplication {
