+++ boot/platform/src/main/java/com/plate/boot/inventory/relational/logger/LoggersController.java
@@ -1,4 +1,4 @@
-package com.plate.boot.relational.logger;
+package com.plate.boot.inventory.relational.logger;
 
 import com.plate.boot.commons.utils.ContextUtils;
 import lombok.RequiredArgsConstructor;
