+++ boot/platform/src/main/java/com/plate/boot/inventory/relational/menus/MenusService.java
@@ -1,4 +1,4 @@
-package com.plate.boot.relational.menus;
+package com.plate.boot.inventory.relational.menus;
 
 
 import com.plate.boot.commons.base.AbstractDatabase;
