+++ boot/platform/src/test/java/com/plate/boot/ApplicationTests.java
@@ -5,10 +5,8 @@ import org.junit.jupiter.api.Test;
 /**
  * @author <a href="https://github.com/vnobo">Alex Bob</a>
  */
-//@SpringBootTest
-class BootApplicationTest {
-
+class ApplicationTests {
     @Test
-    void contextLoads() {
+    void writeDocumentationSnippets() {
     }
 }
