+++ build.gradle
@@ -4,10 +4,6 @@ plugins {
     id 'org.graalvm.buildtools.native' version '0.10.4'
 }
 
-ext {
-    set('springModulithVersion', "1.3.1")
-}
-
 configure(allprojects) { project ->
 
     group "com.alex.plate"
@@ -72,11 +68,5 @@ configure(subprojects - project(":boot")) { project ->
 
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    }
-
-    dependencyManagement {
-        imports {
-            mavenBom "org.springframework.modulith:spring-modulith-bom:${springModulithVersion}"
-        }
     }
 }
